+++ collector/src/main/java/com/bitmovin/analytics/data/Backend.java
@@ -1,6 +1,0 @@
-package com.bitmovin.analytics.data;
-
-public interface Backend {
-    void send(EventData eventData);
-    void sendAd(AdEventData eventData);
-}